created searchBar component and added basic CSS
