+++ src/components/molecules/NavButtons.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import styled from 'styled-components'
 import { Link } from 'gatsby'
-
 import { DropdownMenu, ShoppingBagIcon } from '../atoms'
+import Search from './SearchBar'
 
 const Container = styled.div`
   display: flex;
@@ -80,12 +80,25 @@ const Container = styled.div`
 `
 
 class NavButtons extends Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      showPopup: false,
+    }
+    this.togglePopup = this.togglePopup.bind(this)
+  }
+
+  togglePopup() {
+    console.log(this.state)
+    this.setState({ showPopup: !this.state.showPopup })
+  }
   render() {
     //Get user links and help links that are passed down as props from NavBar - come from contentful
     const userLinks = this.props.userLinks[0].dropdownLinks
     const helpLinks = this.props.helpLinks[0].dropdownLinks
     return (
       <Container>
+        {this.state.showPopup ? <Search /> : null}
         <div className="leftNav">
           <img
             style={{
@@ -95,6 +108,7 @@ class NavButtons extends Component {
             }}
             src={this.props.searchIcon}
             alt="search-icon"
+            onClick={this.togglePopup}
           />
         </div>
         <div className="logo">
